backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
groq>=1.1.2
sentence-transformers==3.3.1
faiss-cpu>=1.9.0
pdfminer.six==20231228
python-dotenv==1.0.1
numpy==1.26.4
